+++ ios/ReactNativeSynerise/Modules/RNContent.h
@@ -1,0 +1,17 @@
+//
+//  RNContent.h
+//  ReactNativeSynerise
+//
+//  Created by Krzysztof Kurzawa on 17/04/2020.
+//  Copyright © 2020 Facebook. All rights reserved.
+//
+
+#import "RNBaseModule.h"
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface RNContent : RNBaseModule <RCTBridgeModule>
+
+@end
+
+NS_ASSUME_NONNULL_END