+++ lib/classes/api_queries/DocumentsApiQuery.d.ts
@@ -1,0 +1,13 @@
+import { DocumentsApiQueryType } from "./DocumentsApiQueryType";
+declare class DocumentsApiQuery {
+    type: DocumentsApiQueryType;
+    typeValue: String;
+    version: String;
+    constructor(type: DocumentsApiQueryType, typeValue: String, version: String);
+    toObject(): {
+        type: DocumentsApiQueryType;
+        typeValue: String;
+        version: String;
+    };
+}
+export { DocumentsApiQuery };