+++ lib/classes/models/Content/RecommendationOptions.d.ts
@@ -1,0 +1,6 @@
+declare class RecommendationOptions {
+    slug: String | undefined;
+    productID: String | undefined;
+    toObject(): object;
+}
+export { RecommendationOptions };